+++ src/main/java/lotto/Application.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.generator.RandomLottoGenerator;
+import lotto.view.LottoGameConsoleView;
+
 public class Application {
     public static void main(String[] args) {
         LottoGame lottoGame = new LottoGame(new LottoGameConsoleView(), new RandomLottoGenerator());
