+++ src/main/java/lotto/constant/LottoConstant.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.constant;
 
 public class LottoConstant {
