+++ src/main/java/lotto/domain/LottoPurchase.java
@@ -1,7 +1,9 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.constant.LottoConstant;
+import lotto.generator.LottoGenerator;
 
 public class LottoPurchase {
