+++ src/main/java/lotto/domain/LottoWinningNumber.java
@@ -1,7 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.constant.LottoConstant;
 
 public class LottoWinningNumber {
