--- src/main/java/lotto/dto/LottoWinningStatisticsResponse.java
@@ -1,4 +1,0 @@
-package lotto.dto;
-
-public class LottoWinningStatisticsResponse {
-}
