+++ src/main/java/lotto/generator/LottoGenerator.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.generator;
+
+import lotto.domain.Lotto;
 
 public interface LottoGenerator {
