+++ src/test/java/lotto/domain/LottoWinningNumberTest.java
@@ -1,10 +1,11 @@
-package lotto;
+package lotto.domain;
 
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import lotto.constant.LottoConstant;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
