+++ src/test/java/lotto/generator/RandomLottoGeneratorTest.java
@@ -1,8 +1,9 @@
-package lotto;
+package lotto.generator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
